Save forecast from searchList and display it in forecast adapter

+++ app/src/main/java/com/mbojec/halo/database/ForecastDao.kt
@@ -15,7 +15,7 @@ interface ForecastDao{
     fun loadSimpleForecast(): LiveData<ForecastEntity>
 
     @Query("SELECT * FROM forecast WHERE cityId = :id")
-    fun loadForecast(id: Int): LiveData<ForecastEntity>
+    fun loadForecast(id: Long): LiveData<ForecastEntity>
 
     @Query("SELECT * FROM forecast")
     fun loadAllForecasts(): LiveData<List<ForecastEntity>>
